add mechanize to banner + refactor css
