+++ src/components/App/content.tsx
@@ -8,7 +8,7 @@ import ErrorBoundary from '../ErrorBoundary';
 import HeaderDisplay from '../HeaderDisplay';
 import Map from '../Map';
 import Finals from '../Finals';
-import SurveyBanner from '../SurveyBanner';
+import SponsorBanner from '../SponsorBanner';
 import {
   AppNavigationContext,
   AppMobileNav,
@@ -40,7 +40,7 @@ function AppContentBase(): React.ReactElement {
         tabs={NAV_TABS}
         captureRef={captureRef}
       />
-      <SurveyBanner />
+      <SponsorBanner />
       <ErrorBoundary
         // ErrorBoundary.fallback is a normal render prop, not a component.
         // eslint-disable-next-line react/no-unstable-nested-components
